common/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "None"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "datamodel-code-generator (>=0.31.2,<0.32.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
